py_vector/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py_vector"
version = "0.0.1"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { file = "LICENSE" }  # Ensure you have a LICENSE file
readme = "README.md"  # Ensure you have a README.md file
requires-python = ">=3.8"

# Dependencies
dependencies = [
    "requests",
    "eth-brownie",
    "pandas",
    "autopep8",
    "isort",
    "pyyaml",
    "beautifulsoup4",
    "omegaconf",
    "pydantic",
    "python-dotenv",
    "lxml",
]

# Optional features (optional)
[project.optional-dependencies]
dev = ["pytest", "black"]  # Add any development dependencies


[tool.setuptools.packages.find]
where = ["."]
include = ["py_vector*"]  # Include only the `py_vector` package
